refactor: migrate grant-calendar page to feature architecture
- Extract GrantCalendarContainer into src/features/grant-calendar/
- Replace AdminContainer with AppPageLayout + FloatingActions
- Replace DeleteConfirmationModal with Dialog component
- Add tabs support to AppPageLayout
- Mark AdminContainer as deprecated
