+++ src/features/grant-calendar/components/index.ts
@@ -1,0 +1,1 @@
+export {GrantCalendarContainer} from './GrantCalendarContainer';
