+++ src/shared/components/layout/AppPageLayout.tsx
@@ -2,11 +2,24 @@
 
 import {ReactNode} from 'react';
 
-import {Alert} from '@heroui/react';
+import {Alert} from '@heroui/alert';
+import {Tab, Tabs} from '@heroui/tabs';
 
 import {LoadingSpinner, VStack} from '@/components';
 
 import {commonCopy} from '../../copy';
+
+export interface TabConfig {
+  key: string;
+  title: ReactNode;
+  content: ReactNode;
+
+  actions?: ReactNode;
+  filters?: ReactNode;
+
+  inheritGlobalActions?: boolean;
+  inheritGlobalFilters?: boolean;
+}
 
 export interface AppPageLayoutProps {
   header?: ReactNode;
@@ -17,6 +30,11 @@ export interface AppPageLayoutProps {
   isLoading?: boolean;
   isError?: boolean;
   isUnderConstruction?: boolean;
+
+  tabs?: TabConfig[];
+  activeTab?: string;
+  onTabChange?: (key: string) => void;
+  tabsAriaLabel?: string;
 
   children: ReactNode;
 }
@@ -29,9 +47,31 @@ export function AppPageLayout({
   isLoading,
   isError,
   isUnderConstruction,
+
+  tabs,
+  activeTab,
+  onTabChange,
+  tabsAriaLabel = 'Tabs',
+
   children,
 }: AppPageLayoutProps) {
   if (isLoading) return <LoadingSpinner />;
+
+  const effectiveActiveTab = activeTab || tabs?.[0]?.key;
+
+  const renderTabContent = (tab: TabConfig) => {
+    const tabActions = tab.inheritGlobalActions ? actions : (tab.actions ?? undefined);
+
+    const tabFilters = tab.inheritGlobalFilters ? filters : (tab.filters ?? undefined);
+
+    return (
+      <VStack spacing={4}>
+        {tabActions}
+        {tabFilters}
+        {tab.content}
+      </VStack>
+    );
+  };
 
   return (
     <>
@@ -50,10 +90,30 @@ export function AppPageLayout({
             description={commonCopy.alerts.errorLoadingPage}
           />
         )}
+
         {header && header}
-        {actions && actions}
-        {filters && filters}
-        {children}
+
+        {!tabs && (
+          <>
+            {actions && actions}
+            {filters && filters}
+            {children}
+          </>
+        )}
+
+        {tabs && (
+          <Tabs
+            selectedKey={effectiveActiveTab}
+            onSelectionChange={(key) => onTabChange?.(key as string)}
+            aria-label={tabsAriaLabel}
+          >
+            {tabs.map((tab) => (
+              <Tab key={tab.key} title={tab.title}>
+                {renderTabContent(tab)}
+              </Tab>
+            ))}
+          </Tabs>
+        )}
       </VStack>
 
       {floatingActions}
